broke the build, sorry
stupid me, should know to at least run npm test after
making a "trivial" 2 line change. Of course by changing
default values it made a big impact on our tests which
mostly rely on the default values being honored. I
quickly resolved the issue by explicitly adding the old
values to the tests.

I should go through all of the app tests and see if I can
update any of them back to using the default values.
However, it is late, and we still have the same coverage
with this workaround. So for tonight the tests are back
doing what they were doing before, just in many cases
that's not the default thing.