provide basic and advanced mode for main generator
This is based on suggestion from Ole
which was relayed from others at Alfresco

Basic mode is a very minimal layer on top
of AIO. Adding scripts and a few other
things.

Advanced mode adds the customizations
folder structure to basic mode. This is
how generator-alfresco worked previously.

We decided to default to basic mode in
order for projects to more closely align
with the SDK.

resolves #126